fix(接口): 新搜索接口分数计算优化

+++ housekeeping-common/src/main/java/com/housekeeping/common/entity/ScoreCalculation.java
@@ -58,7 +58,7 @@ public class ScoreCalculation {
     private Float getScope2(){
         Float maxScope = new Float(weight.get(ApplicationConfigConstants.distanceScoreDouble));
         Float x = new Float(this.variable6);
-        Float y = (-1) * maxScope * x/new Float(10000) + maxScope;
+        Float y = (-1) * maxScope * x/new Float(50000) + maxScope;
         return y;
     }
